:bug: (logs): Fix versioning mismatch of hw-transport
