+++ .changeset/polite-pumas-marry.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/logs": minor
+---
+
+Json replacer fix
