+++ package.json
@@ -230,6 +230,7 @@
       "remove-flow-types-loader>flow-remove-types": "^2",
       "remove-flow-types-loader>loader-utils": "*",
       "@ledgerhq/devices": "workspace:*",
+      "@ledgerhq/hw-transport": "workspace:*",
       "tslib": "2.6.2",
       "@hashgraph/sdk>@hashgraph/cryptography": "1.1.2",
       "@hashgraph/sdk>@grpc/grpc-js": "1.6.7",
